+++ src/main/java/com/develhope/spring/exception/NotFoundException.java
@@ -1,4 +1,4 @@
-package com.develhope.spring.controllers;
+package com.develhope.spring.exception;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.ResponseStatus;
